+++ src/guide/content/component/indeterminate.svelte.ts
@@ -1,0 +1,8 @@
+import type { Action } from "svelte/action"
+
+export const indeterminate: Action<HTMLInputElement, { state?: boolean }> = (
+	node: HTMLInputElement,
+	data,
+) => {
+	if (data.state) node.indeterminate = true
+}